requirements/base.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.11.14
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-disqus==0.5
django-tinymce==2.7.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drfdocs==0.0.11
enum34==1.1.6
funcsigs==1.0.2
idna==2.7
ipaddress==1.0.22
mock==2.0.0
pbr==4.2.0
Pillow==5.2.0
pycparser==2.18
PyJWT==1.6.4
pyOpenSSL==18.0.0
pytz==2018.5
requests==2.19.1
six==1.11.0
stripe==2.5.0
urllib3==1.23
gunicorn==19.4
whitenoise==3.2